tests/examples/sample_web_test.py
"""
Web UI测试样例
演示如何使用框架进行Web UI测试
"""

import pytest
import allure
from page_objects.login_page import LoginPage
from utilities.logger import log


@allure.epic("Web UI测试样例")
@allure.feature("登录功能")
class TestSampleLogin:
    """登录功能测试样例"""
    
    @allure.story("成功登录")
    @allure.severity(allure.severity_level.CRITICAL)
    @pytest.mark.web
    @pytest.mark.sample
    def test_sample_login_success(self, web_driver, web_config):
        """测试成功登录样例"""
        with allure.step("演示Web UI测试框架功能"):
            # 这是一个示例测试，展示如何使用框架
            # 在实际测试中，这里会进行真实的Web UI测试
            log.info("Web UI测试框架功能演示")
            
            # 模拟页面对象使用
            log.info("创建登录页面对象")
            log.info("导航到登录页面")
            log.info("输入用户名和密码")
            log.info("点击登录按钮")
            log.info("验证登录结果")
            
            # 模拟截图
            log.info("截图记录测试过程")
            
            # 模拟断言
            assert True, "示例测试通过"
