requirements-dev.txt
# 开发环境专用依赖包

# 调试工具
debugpy>=1.8.0
ipdb>=0.13.13
pdb++>=0.10.3

# 测试工具
pytest-xdist>=3.3.1
pytest-cov>=4.1.0
pytest-html>=3.2.0
pytest-mock>=3.11.1
pytest-benchmark>=4.0.0
pytest-timeout>=2.1.0
pytest-rerunfailures>=12.0
pytest-sugar>=0.9.7
pytest-clarity>=1.0.1

# 代码质量
flake8>=6.0.0
pylint>=2.17.5
black>=23.7.0
isort>=5.12.0
bandit>=1.7.5
safety>=2.3.5
mypy>=1.5.1
pre-commit>=3.3.3

# 代码复杂度分析
radon>=6.0.1
xenon>=0.9.0
vulture>=2.9.1
mccabe>=0.7.0

# 文档生成
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0

# 性能分析
memory-profiler>=0.61.0
line-profiler>=4.1.1
py-spy>=0.3.14

# 开发服务器
watchdog>=3.0.0
livereload>=2.6.3

# Jupyter相关
jupyter>=1.0.0
jupyterlab>=4.0.5
notebook>=7.0.2
ipython>=8.14.0
ipywidgets>=8.1.0

# 数据分析（用于测试数据分析）
pandas>=2.0.3
numpy>=1.24.3
matplotlib>=3.7.2
seaborn>=0.12.2

# API开发和测试
fastapi>=0.101.1
uvicorn>=0.23.2
httpx>=0.24.1

# 数据库开发工具
alembic>=1.11.2
sqlalchemy-utils>=0.41.1

# 容器化开发
docker>=6.1.3
docker-compose>=1.29.2

# 环境管理
python-dotenv>=1.0.0
environs>=9.5.0

# 日志增强
colorlog>=6.7.0
structlog>=23.1.0

# 时间处理
freezegun>=1.2.2
python-dateutil>=2.8.2

# 文件处理
pathlib2>=2.3.7
watchfiles>=0.19.0

# 网络工具
requests-mock>=1.11.0
responses>=0.23.3

# 加密和安全
cryptography>=41.0.3
passlib>=1.7.4

# 配置管理
pydantic>=2.1.1
dynaconf>=3.2.0

# 任务队列（开发测试）
celery>=5.3.1
redis>=4.6.0

# 缓存
diskcache>=5.6.3

# 序列化
marshmallow>=3.20.1
jsonschema>=4.19.0

# 命令行工具
click>=8.1.6
rich>=13.5.2
typer>=0.9.0

# 代码格式化插件
flake8-docstrings>=1.7.0
flake8-import-order>=0.18.2
flake8-bugbear>=23.7.10
flake8-comprehensions>=3.14.0

# 类型检查插件
types-requests>=2.31.0.2
types-PyYAML>=6.0.12.11
types-redis>=4.6.0.5

# Git hooks
gitpython>=3.1.32

# 模拟和存根
factory-boy>=3.3.0
faker>=19.3.0

# 并发测试
pytest-asyncio>=0.21.1
aiohttp>=3.8.5

# 图像处理（截图相关）
Pillow>=10.0.0

# 报告生成
jinja2>=3.1.2
weasyprint>=59.0

# 开发工具集成
rope>=1.9.0
autopep8>=2.0.2

# 性能监控
psutil>=5.9.5

# 数据验证
cerberus>=1.3.5
voluptuous>=0.13.1
